[EI-Connect] 1. move svbversion.h into wise for EI-Connect version control 2. modify Sample code comment.
